Assert: Factory
