+++ FL_TESTS
@@ -1,6 +1,7 @@
 test/AssertActivation.h
 test/AssertConstructor.h
 test/AssertDefuzzifier.h
+test/AssertFactory.h
 test/Headers.h
 test/Listener.h
 test/Mock.h
